hack to deal with snowflake including quotes in names (#29)
hack to deal with snowflake including quotes in namesIt appears that at some point snowflake started including double-quotes in role names when you use that syntax to create role grants. It appears to be a regression, since our code used to work and no longer does (with no changes).

This hack should be safe once they fix it because double-quote is not a legal part of the name.