Enable support for externalbrowser Authenticator (#28)
Enable support for externalbrowser AuthenticatorThis PR adds supports the "externalbrowser" authenticator. In this mode the user will be redirected to their browser to authenticate. 

This works well for accounts with SAML, but is a tad silly since it pops a browser tab for each connection. Long term we should figure out how to cache credentials here.