add Exists function to user object
Without this it will just try to read a user and in the case where a user has been deleted, but is in the state file
we will get an error. There may be a graceful way to handle that in the Read function, but this seems like a cleaner way to do it.