Fix a small connection leak
This fixes a connection leak which causes open connections to accumulate
for each schema that gets refreshed.  This probably doesn't manifest as
a user-facing bug, but it certainly can't hurt to merge a fix.

For an explanation on how I discovered and tested this, see the
following github issue comment:

https://github.com/chanzuckerberg/terraform-provider-snowflake/issues/69#issuecomment-528959865