[fix] Fix cannot convert NULL to string when creating stage (#129)
When I try to create an internal stage, Terraform breaks with an error
trying to convert a NULL value to a string, which is not possible in Go.

When you run SHOW STAGES in internal stages, the "region" and "cloud"
fields are NULL. This pull request fixes that by using *string pointers
in the SHOW STAGE struct.