Add option to SKIP_DATABASE_TESTS and SKIP_WAREHOUSE_TESTS
In the Snowflake Standard accouynt, multi-warehouses are not enabled,
which make the warehouse integration tests fail. Similarly, the
database integration tests fail because it doesn't allow retention
days larger than 1 day.

A contributor using a Snowflake trial account to open a pull request
will need those two options make the tests pass in the pull request.
Instructions on how to do it have been included in the README.