[tests] switch testify/assert to testify/require
Replace all useage of the testify/assert library with its equivalent
testify/require.

testify/assert is commonly misunderstood by developers who expect it to
terminate test runs on the first failure. However that is not what it
does. To halt tests on first failure you need to use testify/require.