[fix] statement extraction logic for view resource
Support CTE when extracting queries from view definitions.

When a view definition gets returned from snowflake it has the `create
view ...` at the beginning. To compare that sql statement to our state
we need to strip down to the query statement used to define the contents
of the view. Typically that statement starts with SELECT. When Common
Table Expressions (CTEs) are used WITH can start the statement.

As with some other PRs, I would love to have a proper SQL statement
parser here, but for now, just incrementally make our hacks better.