account: update grants in-place instead of remove/add method
Terraform supports updating a resource in-place if a Terraform schema
key is not marked as `ForceNew` and an `Update` method is provided. In
fact, it is implemented for 2 resources already:

- https://github.com/igungor/terraform-provider-snowflake/blob/e870133b3e788b39caae7f96035c528e8fdb9b44/pkg/resources/schema_grant.go#L90
- https://github.com/igungor/terraform-provider-snowflake/blob/e870133b3e788b39caae7f96035c528e8fdb9b44/pkg/resources/role_grants.go#L18

This PR aims to add in-place support for 'account grants'.

We're testing this method at @peak for a while. This is critical because
'force delete/add' method makes services hiccup while trying to access a
Snowflake resource.

I plan to add more of this for:
- database grants
- table grants
- view grants
- warehouse grants

also update:
- schema grants
- role grants

to have the same structure as the others if this PR looks good to you.

Updates #137