Update procedure resource to set all values during import
Update ReadProcedure() to set name, database and schema from Show during
import.

Remove the RETURN argument from ArgumentsSignature() because it is not
part of the unqiue signature of a procedure and will not exist during
import unless the ID is updated to include it.

Change default value of comment to 'user-defined procedure' and fix
corresponding tests.