Read Database and Schema name during import
Update the stream resource to read both the Database name and the
Schema name along with the other attributes so that after importing an
existing Stream you are not prompted to recreate the Stream due to the
missing values.

Update both the unit test and the acceptance test to verify a
successful read and import.

Addresses Issue 466.