fix: Correctly read INSERT_ONLY mode for streams
This change removes the `append_only`/`insert_only` fields from the
Stream row struct, as thy are not returned neither by `SHOW STREAMS` nor
`DESCRIBE STREAMS`.

To find if an existing stream is `INSERT_ONLY`, it now checks the `mode`
field, as it has been already done for the `APPEND_ONLY` model.