Add client credential caching option
This is normally _disabled_ for Linux on the upstream gosnowflake SDK;
because Linux doesn't offer a password store like Windows and Mac do. I
didn't add platform-detecting behaviour here to avoid surprises; instead
hopefully the docs should indicate whether this is something you'd want
or not.

This must be enabled on the snowflake account. See https://docs.snowflake.com/en/user-guide/admin-security-fed-auth-use#using-connection-caching-to-minimize-the-number-of-prompts-for-authentication-optional

Specifically, as ACCOUNTADMINISTRATOR:

```
alter account set allow_id_token = true;
```