feat: make refresh_mode a settable option
Snowflake recently added the ability to set REFRESH_MODE manually when
creating dynamic tables. Generally this is for use cases where the user
wants to force the DT to use a FULL refresh when the automatic behavior
would otherwise be INCREMENTAL. Some reasons users may want to override
INCREMENTAL in favor of FULL include a) performance, and b) some edge
cases where DTs fail to refresh unless they are forced to use FULL.

This commit updates the terraform dynamic table resource to allow users
to set `refresh_mode` in configuration. This value can only be set when
the DT is created, and forces replacement if the value changes.

We suppress most diffs for this parameter because Snowflake also allows
users to specify `AUTO` as an input value for refresh_mode, which should
be ignored on refresh.