feat: add initialize setting to dynamic table
Dynamic tables have an INITIALIZE parameter that can be specified on
table creation. This parameter determines when the first refresh of the
dynamic table occurs. This option defaults to ON_CREATE, which causes
the dynamic table to execute a refresh immediately upon creation.

There are cases where this behavior is not desirable. For example, if
refreshing takes a long time, we may not want terraform to wait. There
are also some cases where a refresh ON_CREATE fails due to time travel
issues. In these cases, setting the initialize option to ON_SCHEDULE may
be the solution for these issues.

This value does not have its own column in SHOW DYNAMIC TABLES and can
only be parsed from DDL.

This commit also changes the data_timestamp field to a nullable type.
This is necessary because when dynamic tables are set to INITIALIZE =
ON_SCHEDULE, the dynamic table's data_timestamp is null at creation
time, and is only populated when the table refreshes.