fix: allow DT warehouse to be updated in-place
Dynamic table warehouses can be updated in-place with an ALTER DYNAMIC
TABLE statement. The terraform resource was requiring the DT to be
dropped and recreated in order to change the warehouse, but this is not
necessary.

This commit removes the ForceNew modifier on the "warehouse" parameter
so that users can easily change the warehouse. This is a setting that
users may want to easily change for cost allocation purposes.