fix: Fix tests that base on default data retention (#2465)
There were some tests failing on the CI since Thursday/Friday (2 in
application packages, multiple in application and one for table
cloning). The same tests were passing "locally". The reason for all of
them was the assumption that the account level parameter
`DATA_RETENTION_TIME_IN_DAYS` is set to `1` (default). For some reason,
on our CI account, it was changed to `0` (either by some test in which
cleanup failed or manually by someone). I did two things so that our
tests pass again:
- reverted the param to `1` default
- prepared a fix for the tests, as to not make assumptions about this
param (or for table cloning setting it temporarily to `1` and reverting
to the previous value after the test)