fix: Fix task related issues (#2479)
Related issues: #2036 #2207 #2346.

- (#2207) The reason was the change in parsing the `SHOW TASKS` response
for `after` parameter. In earlier versions, the parsing was incorrect
for the identifiers containing dot characters. The new one fixed the dot
case but Snowflake identifiers surprised us one more time, and the basic
case, without any `"` was not supported correctly.
- (#2036) There is no way to unset `when` currently in Snowflake. We
don't want to always `ForceNew` for when change. For that reason we run
forceNew conditionally using `customdiff.ForceNewIfChange`.
- (#2346) Waiting for answers but maybe this fix will also solve this
issues.