Nake DT query parser handle two comment positions
The "correct" place to specify the query comment in a dynamic table DDL
is at the end of the metadata argument list, right before the query.
However, when retrieving an existing query from Snowflake using SHOW
DYNAMIC TABLE, the comment is placed at the beginning of the metadata
argument list.

This commit updates the DT query parser to handle comments that can show
up in either position.