fix: parse dynamic table query from DDL (#2438)
Previously we have been extracting the `query` portion of dynamic table
definitions from the stored DDL in Snowflake by searching for the string
position of keywords that preceded the query. However, this approach has
proven inadequate as there are many ways to write working dynamic table
definitions that conflict with string position assumptions.
    
Observing that definitions of views and materialized views have faced
similar challenges but are currently working, this commit extends the
existing ViewSelectStatementExtractor that is currently being used to
extract the query definition from view and materialized view DDL
statements, applying the same approach to dynamic table DDL statements.

This is an attempt to address #2329

## Test Plan
* [x] acceptance tests
* [x] testing in our existing stack. We have an existing stack with ~100
dynamic tables where many have CTEs and fail to be parsed with v0.85.0.
A few contains `AS` in the name and were affected by spurious diffs with
v0.78.0 (as observed in #2329). All of our dynamic tables are properly
preserved with code change in this PR.

## References

Original issue:
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/2329
Previous attempt to fix:
https://github.com/Snowflake-Labs/terraform-provider-snowflake/pull/2421

---------

Co-authored-by: Artur Sawicki <artur.sawicki@snowflake.com>