fix: allow DT warehouse to be updated in-place (#2439)
Dynamic table warehouses can be updated in-place with an ALTER DYNAMIC
TABLE statement. The terraform resource was requiring the DT to be
dropped and recreated in order to change the warehouse, but this is not
necessary.

This commit removes the ForceNew modifier on the "warehouse" parameter
so that users can easily change the warehouse. This is a setting that
users may want to easily change for cost allocation purposes.

## Test Plan
* [ ] tested by running `terraform apply` and `terraform plan` on an
existing stack with warehouse changes, observed that dynamic tables were
updated in place, and `ALTER DYNAMIC TABLE` was run in Snowflake

## References

Documentation for `ALTER DYNAMIC TABLE` command:
https://docs.snowflake.com/en/sql-reference/sql/alter-dynamic-table

---------

Co-authored-by: Artur Sawicki <artur.sawicki@snowflake.com>