fix: user error handling (#2486)
Fix for
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/2481
Adjusted error handling in the User's Read operation. Now when an error
happens and we get the "object does not exist or not authorized" error
then we're marking the resource as removed and return nil. Otherwise, we
return the error. Also, added a test case that proves no error happens
when the user is removed outside of the Terraform (it failed with the
previous implementation).