fix: SHOW GRANTS mapping for share data type (#2508)
Fixes:
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/2410
The root cause of the issue was that in SHOW GRANTS conversion from row
representation to SDK object representation, we had a special case for
shares that would split it on dot. Share may have only one part and this
change adds support for such a situation. Additionally, it was covered
by an existing test that was commented out and was waiting for the
`snowflake_grant_privileges_to_share` resource.