fix: data retention time parameters follow-up (#2530)
<!-- Feel free to delete comments as you fill this in -->
A follow-up for
https://github.com/Snowflake-Labs/terraform-provider-snowflake/pull/2502.
In this pr I applied the same changes to the Database and Table
resource. On the note side, I left some comments in the table resource.
There's still a deprecated data retention parameter that I would remove
because it wouldn't be used with the new implementation that always
expects a value (-1 if not set). We can add a note in the migration note
and change the description of the deprecated parameter stating that it
doesn't have any effect on the table configuration. Also in the database
tests I'm changing account-level data retention time which I think
shouldn't (but may) affect other tests. I would leave it anyway, because
on delete we set back the default value and crashes may occur, but very
rarely (I think).

## Test Plan
<!-- detail ways in which this PR has been tested or needs to be tested
-->
* [x] acceptance tests
<!-- add more below if you think they are relevant -->