fix: grant read operation (#2665)
This pr resolves an issue that was happening when a granted resource was
removed by hand. This touches all the `grant privilege to` resources
(others too, but they somewhat already resolved this by marking the
resource as removed; the other resources are only missing nested `if`
check that would indicate for which error we're making the resources as
removed. Right now, it's happening for every kind of error in the `SHOW`
command). I adjusted those resource to:
1. Check if the granted resource is present (e.g. grant x to database
role -> in Read operation check if database role is present).
2. Check the error type in the `SHOW GRANTS`, so we'll only mark the
resource as removed only for `object does not exist or not authorized`
type of errors.

## Test Plan
<!-- detail ways in which this PR has been tested or needs to be tested
-->
* [x] acceptance tests that show resources are marked as removed without
dependent resources (grant privileges to share have only one test
because Snowflake is preventing you from removing dependent objects like
databases when they're granted to share)