feat: grant ownership follow-up (#2658)
A follow-up for
https://github.com/Snowflake-Labs/terraform-provider-snowflake/pull/2604.

Done in this pr:
- All of the edge cases handled and tested (except of tasks that are
done in the separate pr):
  - Materialized views (already handled by Snowflake no changes needed)
  - RBAC hierarchy (test case added)
- Delete dependent resource (role or granted object) and remove grant
resource from the state (test case added)

Won't do:
- External tables (cannot handle this edge case, because we have to know
the auto_refresh state of the external table; it's not retrievable by
SHOW or DESC commands. It will be still possible to grant ownership of
the external table, but there may be additional manual work to do
afterward. Everything is documented.)

## Test Plan
<!-- detail ways in which this PR has been tested or needs to be tested
-->
* [x] acceptance tests that show how the resource is handling certain
edge cases + RBAC use case

## References
[GRANT
OWNERSHIP](https://docs.snowflake.com/en/sql-reference/sql/grant-ownership)

## Mentioned in
A list of issues requesting this resource:
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/2549
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/2199
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/2084
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/1942
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/1875