fix: renames in update operations (#2702)
This change fixes all of the Updates that were performing re-names to
always setId at the end of the rename operation as well as updating id
that was used in the later operations. Also, the `ForceNew`s were
removed in cases where rename was also handled in the Update. Test cases
were added to show that rename is going through Update and not
DeleteAndCreate operation + additional param is always changed to see if
other operations are affected by the previous rename operation. The
table constraint was adjusted and cannot support the rename for now
(until the Read operation is working #2683). Migration guide updated for
behavioral changes.

---------

Co-authored-by: Artur Sawicki <artur.sawicki@snowflake.com>