chore: Update the contribution guidelines (and small fixes) (#2753)
The main part of the change is refreshing the outdated contribution
guidelines. Note, that part on advanced debugging was just improved with
the building instructions and will be validated separately (because it
was updated not that long ago). Keep in mind, that this can be an
iterative process, we can still improve them but we should have a good,
refreshed based already.

Additionally, the TODO comments were updated and enriched with the
proper ticket references.