fix: Invoke SetId immediately after alert creation (#2786)
Now, after failing alert alter operation, terraform keeps the created
alert in state. A warning is shown to a user when alter fails. Alert can
be now enabled by using terraform apply.

Also, CRUD functions are changed to functions with context due to
deprecations in terraform sdk.

This change was tested on a real Snowflake environment. 
<!-- Feel free to delete comments as you fill this in -->

<!-- summary of changes -->

## Test Plan
<!-- detail ways in which this PR has been tested or needs to be tested
-->
* [ ] acceptance tests
<!-- add more below if you think they are relevant -->
* [ ] …

## References
<!-- issues documentation links, etc  -->

*