chore: Replace parsing function for saving granted object names (#2813)
Add better parsing of identifiers returned by SHOW GRANTS. It should
resolve some of our issues with failing tests (they sometimes fail
because they depend on the generated name of a resource and whether they
will be quoted in the output provided by Snowflake).

New `ParseObjectIdentifier` function is based on
`DecodeSnowflakeParameterID` function used in resources. As a fallback
(because `ParseObjectIdentifier` can fail)
`NewObjectIdentifierFromFullyQualifiedName` will be used.