chore: accept non-pointer values in the generated builder methods (#2816)
Accept non-pointer values in the generated builder methods. Now the
nested requests have to be explicitly unreferenced, but two additional
cases could be added to resolve this "issue". There's only a special
case for strings which shouldn't be set whenever they're optional and
the input value is equal to an empty string, because:
1. that would be a rare case to have an empty string in Snowflake.
2. empty strings are treated as non-set values in Terraform.

The improved generator was tested and used as an example on the external
table.