feat: standard database v1 readiness (#2842)
Done in this pr:
- Addressed comments from the [previous
pr](https://github.com/Snowflake-Labs/terraform-provider-snowflake/pull/2834)
- Added new `standard_database` resource (also added examples, import,
mentioned in migration notes)
- Deprecate old database resource
- Adjust `databases` data-source to align with v1 requirements (add
filters, missing values and outputs from DESCRIBE and SHOW PARAMETERS
commands)
- Replaced `snowflake_database` in all of the examples to
`snowflake_standard_database`

To be done:
- Add missing properties on all three new database types
- Make sure all of the issues were resolved with the new types of
databases

## Test Plan
<!-- detail ways in which this PR has been tested or needs to be tested
-->
* [x] acceptance tests

## References
<!-- issues documentation links, etc  -->

[CREATE
DATABASE](https://docs.snowflake.com/en/sql-reference/sql/create-database)

## Update
Changes done:
- Added missing parameters to all the database types and moved them and
operations on them to a common place (only metric_level parameter wasn't
included as it is a preview feature and there wasn't enough information
about and it seemed like ORGADMIN or certain privileges were required to
be able to test/use it).
- Switched to plain values instead of nested ones for parameters and
adjusted customdiffs, so that the state is refreshed always when
expected.
- Every database type resolves database-connected issues (most if not
all of them were already resolved by the latest versions of the provider
for the `snowflake_database` resource).
- Refresh for secondary was not added, as replication guidelines are
recommending to create a task that would be refreshing the replica at a
certain interval. To aid our users I created an example showing how to
create a task that would run the refresh every 10 minutes. An easy
upgrade (if we would like to) would be add a toggle to call refresh
every read operation. The toggle could be turned on by default with the
option to turn it off and refresh it in the task "manually".
- State upgrader for snowflake_database (because we chose to rename the
old one for have _old suffix).