feat: Add identifier parsers (#2957)
Changes:
- Add and test identifier parsers for object and account identifiers
with different separators (only dot was added as identifier parser,
description provided below).

The encoder/parser for resource identifiers was in the end implemented
as simple string join/split by pipe, because of the following:
- Identifiers that have only one part that is represented by the
`sdk.ObjectIdentifier` should be encoded as `FullyQualifiedName` and
parsed with one of the newly introduced parsers. This should improve a
bit the support for tools that generate import commands (or import
blocks) because the resource identifier is the same as the Snowflake
one.
- More complicated identifiers having more parts should use the provided
encoder/parser to create or parse resource identifiers into parts. This
could cause problems for things that may contain pipe, but we should
generally avoid it or document it (can be done in the usage part of
identifier rework). Also, CSV reader/writer cannot be used for resource
identifiers, because this would make identifiers more complex. For
example, a fully qualified identifier with quotes would have to be
imported as `terraform import snowflake_table.test_table
'"""database_name"".""schema_name"".table_name"""'`. To have this as an
identifier would also mean creating a state upgrader for every resource
(or general one that could be reused) that would transform the
identifiers into CSV-compliant ones.

Notes: 
- Implementation with the use of `identifeirParsingFunc` was added
because initially, I had two parsers (one with dot as csv.Comma and the
other one with pipe) and later on removed the one with the pipe as
csv.Comma (I can merge the unexported parsers with exported ones or
leave it as is if we would like to introduce another internal parser).
- ExternalObjectIdentifier and AccountIdentifier parsers were created
with certain assumptions (described in code).

## Test Plan
<!-- detail ways in which this PR has been tested or needs to be tested
-->
* [x] unit tests

## References
<!-- issues documentation links, etc  -->

* SNOW-1495051