fix: database tests and introduce a new parameter (#2981)
## Changes
- #2978 (adjust migration guide)
- Fix database tests (remove tests for state upgraders, because they
were incorrect and document step-by-step process for manual testing; I
went through all of it)
- Introduce a new parameter to address #2826 and test it
- added a drop with retries just to be sure, but I don't think it should
ever fail other than some kind of strange networking error between
database create and drop schema
- I opted to drop the schema after setting id, so if we manage to create
a database, but not drop the schema, the provider will print out a
warning message that the public schema couldn't be dropped and it should
be done manually. The other approaches I considered:
- Do it before setting ID, but that would mean the Terraform running
create again and failing on CREATE DATABASE (because it's already
there).
- Do it like right now, but make sure the public schema is gone in the
alter, but I didn't go for that, because it could lead to further
complications with (what is considered "after create"; after successful
create? how would we keep the state of create status? etc.).

---------

Co-authored-by: Jakub Michalak <jakub.michalak@snowflake.com>