fix: Fix user resource import (#3181)
Handle user import correctly.

Password is empty after the `snowflake_user` import; we can't read it
from the config or from Snowflake.
During the next terraform plan+apply it's updated to the "same" value.
It results in an error on Snowflake side: `New password rejected by
current password policy. Reason: 'PRIOR_USE'.`
The error will be ignored on the provider side (after all, it means that
the password in state is the same as on Snowflake side). Still,
plan+apply is needed after importing user.