feat: Account v1 readiness resource (#3252)
Introduces refactor to already existing `snowflake_account` resource.

### Changes

- Refactored existing account resource
- Completely rewritten acceptance tests
- Introduced a return type from Create command that is used to parse
org_name + acc_name that is later used in creating new resource id. The
output is always returned. Adding another SQL calls to get this
information may an issue. For example, when account was successfully
created, but e.g. current organization_name couldn't be fetched; this
situation will end up with created account, but with no connection to
the terraform config of this account. This could be fixed by manually
importing the resource. We can discuss this on Monday.

### Next pr

- data source
- next account resource