chore: Generate marshal json for each model (#3307)
- generate marshal and `depends_on` for each resource and data source
model
- replace deprecated `FromModel` with the new one utilizing the
aforementioned marshaling
- fix test with explicit empty list (and add it to special variables)
- fix tests with explicit null value (and add it to special variables)
- generalize multiline handling and test it
- use multiline strings for definitions in functions and procedures
config builders
- simplify function and procedure definitions
- format whitespace in function and procedure definitions

For the next PRs (added to the issue):
- current config building for functions/procedures (especially
definitions) is complex:
  - definition is build using multiline golang strong and sprintf
  - whitespace is formatted
- this is added in the config builder wrapped in special multiline
marker
  - hcl-compatible json is generated
  - json converted to hcl
  - hcl formatted using custom formatters
This should be refined and made simpler (e.g. newline replacement,
encoding problems, etc.)

- marking in config builders generators which fields should be handled
as multiline by default (because now e.g. SQL function requires the
definition, so it's generated in basic builder as a string input, and
not the multiline one; because of that it has to be added again with
`WithFunctionDefinitionValue(...)` which is confusing)