fix: Propose changes to assertions setup with varying test client (#3406)
Currently, the Snowflake object assertions have a fixed test client
reference, e.g.:
https://github.com/Snowflake-Labs/terraform-provider-snowflake/blob/9be21966ca5b1187d5c055026dd5f563900a1aa9/pkg/acceptance/bettertestspoc/assert/objectassert/user_snowflake_gen.go#L23.

Because of that, running the SDK integration tests depends on the
acceptance package which is not ideal.

The current PR introduces a proposal on how the test client can be
pushed to the assertions setup:
1. We modify the common assertions setup and helper functions to allow
receiving the test client as an argument. For now, these are just
duplicates that will be cleaned up when we apply them to all existing
object assertions and assertion types.
2. When setting up a specific assertion, instead of passing the provider
directly, we pass a function that receives the test client and returns
the provider.
3. We add a dedicated function for acceptance and integration tests that
takes care of the client pick; this way, we are sure that the developer
picks the proper client (later, it should be also checked on the
architecture tests level - proper importing).

Next steps (1 or more PRs):
1. Apply this for all assertion types that run the VerifyAll method
underneath.
2. Alter the generation to generate default object assert instructors
without specified test client but with the function.
3. Use proper assertions in all tests.
4. Remove deprecated logic (old snowflake object assertion constructor
receiving a provider).