fix: account for null comment (#3417)
## Changes
Fixes null comment case in the account resource described in
https://github.com/Snowflake-Labs/terraform-provider-snowflake/issues/3402.
Because of the limitation in the terraform testing framework, I couldn't
create an account externally and then import the created account as a
first step of the test (because the testing framework seems to be only
working on already existing objects that can be found in-state). That's
why I had to test this change manually by using `make install-tf` and
using a locally built provider. I confirm that before accounts created
with explicitly null comment `COMMENT = null` were failing and after
applying this change they are not failing anymore.