chore: fix failing tests (#3428)
## Changes
Some of the tests may fail after adjusting network policies due to their
connection with account parameters. This PR was made to check which ones
fail after the adjustment and fix them.

**Update:** No tests regarding network policy failed after the change,
but still some of them did fail and I adjusted them in this pr.
- authentication policy changes for show command (it seems the filtering
in show command is fixed, but the documentation still needs adjustments)
- oauth describe output change (couldn't find any BCR for it)
- account acceptance tests failing due to HCL error (multiline string
with `"` instead of heredoc)